Enable two-pass/deferred execution
The command group functors are now called twice: Once during a so-called
"prepass", in which only the buffer accessors as well as the kernel name (for
debugging) are extracted. This information can then used to schedule (NYI) the
exeuction of the kernels in an optimal way. To execute the kernels, the command
group functors are called a second time, now within a proper SYCL execution
context.

The difficulty here lies in finding a way of calling the functors during
prepass, without objects such as the SYCL command group handler available,
which would be required to construct a cl::sycl::accessor - all while not
breaking the single-source kernel compilation.

This is the last major hurdle that had to be shown as workable for this proof
of concept (in regard to interoperability with ComputeCpp).