Compute command graph in sibling set batches
A sibling set is a set of tasks with a common set of requirements. While not
every pair of tasks in a sibling set must share one or more parents, all
siblings are related through some transitive child->parent->child->[...] chain.

This should allow us some freedom when juggling with task distribution over
nodes, while also making it easier to find locations for inserting
"wait-for-pull" commands.