Generate "wait-for-pull" commands
These commands are generated for nodes which contain buffer ranges that are
needed by other nodes (pull). If the owner node writes to that buffer range
within one of the sibling tasks currently being processed, the "wait-for-pull"
command is inserted before the write (compute command). If no such write exists,
the "wait-for-pull" is inserted before the next compute command within the same
task as the corresponding pull.

Note that in the first case the pull/wait-for-pull pair thus spans two separate
tasks!

Also note that we for now assume that every node has a compute command within
every task.