Add new "master access" feature, refactor tasks
This adds a new celerity::with_master_access function to the public API. This
function allows the master node to gather arbitrary buffer ranges for reading
and/or writing, which can e.g. be useful for IO or checking results.

To integrate this feature neatly into the existing command generation pipeline,
it is also modeled as a "task". To accomodate the differences in behavior (e.g.
no "range mapping", as no splitting takes place), the whole idea of tasks is
abstracted and refactored in a major way. From this, some changes to command
graph generation are required as well.

In addition to tasks, prepass/livepass handlers, buffers and accessors are
reorganized to work cleanly with both types of tasks.

During the implementation, a race condition in the currently employed PULL /
AWAIT PULL model has been discovered. Since an actual fix for this requries some
additional work, a temporary workaround is implmented. See
send_job::WORKAROUND_avoid_race_condition for more information.