Ensure buffer data remains valid for async send duration
This fixes two issues that were surfaces by running MPI over a TCP
transport (as opposed to the default shared memory, which is much
faster): MPI_Isend requires that the source data buffer remains valid
until the operation is completed. This was not ensured for either
sending commands or buffers.

For command packages, we now simply use blocking send, as they are small enough.
For buffers, we keep the SYCL accessor around for the entire send
duration.