Distinguish kernel and buffer dimensions for range mapper functions
Previously, range mappers simply mapped from subrange<N> to subrange<N>,
where N is the dimensionality of the buffer on which get_access() is
called. This is incorrect behavior, as the "chunks" for which we need to
dermine buffer range requirements have the same dimensionality as the
kernel from which they were split out of.

This new implementation correctly handles this use-case and tries to
give some helpful error messages to the user along the way. For the
buffer dimensionality (i.e. the return type of the range mapper
function) this can be done during compile time, while the kernel
dimensionality is only known at a later point in time (during runtime).