Rework data transfer model to use PUSH instead of PULL
First and foremost, this allows us to get rid of the hacky
race-condition workaround that was in place for send_jobs, which
effectively introduced a hard synchronization point across worker nodes.
Since there is no longer the need for the PULL request / response dance,
the send_job is removed entirely.

Command graph generation is also simplified. We no longer generate the
commands in batches of "sibling sets", as this can cause deadlocks with
certain types of task configurations. Instead, commands are generated
for each task separately, one by one. All types of inter-task
optimizations and dependencies should be handled through graph
transformations (NYI).

Note that the type of race condition that originally gave rise to the
sibling set idea can now happen again, however is very unlikely to occur
in practice.  This will likely be handled with a dedicated graph
transformation in the future.

The PUSH model greatly simplifies the transfer of buffers, as we no
longer need to keep track of requests (PULLs) and responses (incoming
transfers).  Instead, we simply accept any incoming transfers and try to
match them with AWAIT PUSHes when we actually need them.

This also introduces a new command_id, which is unique per command, and
allows to e.g. easily match an incoming PUSH to an AWAIT_PUSH, by
storing the command id of the awaited push (instead of trying to match
task id, buffer id and ranges, as was done previously). This also allows
to PUSH the same buffer to the same node repeatedly (e.g. in a looped
task), without risking a mix-up.