Move to multithreaded scheduling and execution
This is a major rework that basically guts the runtime and distr_queue
classes and splits their previous jobs into several more decoupled and
multi-threaded helper classes.

The task_manager now does most of the things distr_queue used to do,
i.e. managing the task graph, creating new tasks, etc. It also acts as
the main synchronization point between the scheduler and the main
thread.

The scheduler, currently in its most basic form, decides when to process
tasks into the command graph, and when to flush commands to worker
nodes (currently both are done as soon as a new task is available). This
only exists on the master node.

The executor handles the polling for incoming commands and data
transfers (via the buffer_transfer_manager), as well as managing the job
queue.