Completely overhaul task and command dependencies
This constitutes one of the largest reworks so far. In order to reduce
the potential for race conditions on the task graph as well as command
graph level (some of which were documented and worked around in the
past), the way how the task graph and command graph are generated is
completely reworked.

We now generate anti-dependencies whenever a task or command tries to
write to a buffer region that might simultaneously be in use by another
task / command.

Commands are now transferred to worker nodes including a list of their
dependencies (which includes anti-dependencies), making the whole
execution job system much simpler.

This also adds support for all "cl::sycl::access::mode"s, not just read
and write, as we had before. These now also have correct semantics in
regards to retaining previous buffer contents (something which "write"
previously did NOT do).

Finally, this adds a metric ton of new tests for various parts of the
runtime, and especially for task graph and command graph generation.

Detailed notes:
- Add support for all access modes in task graph, command graph and
  host_accessor.
- Compute anti-dependencies on task and command level.
- Compute task dependencies with subrange precision.
- Handle host-initialized buffers explicitly within the task graph.
- Correctly handle intra-task race conditions.
- Test suites can now configure whether SYCL kernels should be compiled
  or not (using the COMPILE_SYCL_KERNELS define).
- Add the ability to print graphs in graph_generation_tests by passing
  the --print-graphs CLI flag.
- Compile tests with the /bigobj flag for MSVC