Add support for hipSYCL
This adds support for another SYCL implementation besides ComputeCpp: hipSYCL.
hipSYCL is great on Nvidia hardware, as it basically compiles down to plain
CUDA.

ComputeCpp is still used by default, if it is found by CMake. Otherwise it is
assumed that syclcc (hipSYCL's compiler wrapper) is used for the compilation.

The workarounds required for ComputeCpp are now only enabled when targeting
ComputeCpp. In addition, some workarounds are still required for a few quirks
in hipSYCL. In particular, hipSYCL currently does not support device profiling
(however nvprof can be used in the meantime).

The matmul example is changed to do indentity-multiplication, as that is much
faster to verify and does not suffer from implementation dependant floating
point rounding errors.