Refactor CMake setup and installation
Apart from various minor improvements and removal of some dead code,
this mainly changes the way the two supported SYCL implementations are
handled.

First, the newly supported find_package(hipSYCL) is used, making it much
more convenient to use hipSYCL. Due to this, hipSYCL is now the default
SYCL implementation that Celerity checks for upon configuration. If it
cannot be found, it falls back to ComputeCpp.

Secondly, the chosen SYCL implementation is now recorded to the
installed CMake package configuration file, ensuring that the same
SYCL implementation is used during the installation as well as
subsequent usage of the library.

Due to a technicality (restricting hipSYCL device header paths), the
installed include directory structure needs to be adjusted as well.
However, this new approach is actually somewhat cleaner, as it doesn't
pollute the root include directory with headers that are essentially
implementation details.