Clear up various lifetime issues
This is a bigger overhaul of various internals as well as public APIs in
order to clear up issues regarding the lifetime of Celerity runtime
objects.

The main issue turned out to be the fact that the runtime only waits for
the completion of all tasks once the distr_queue goes out of scope.
However, the distr_queue is typically allocated before any buffers on
the stack, which means that all buffers go out of scope before
distr_queue once the stack unwindws. This in turn leads to all buffer
references that have been captured within command group functions to
become stale.

To avoid this, while retaining the overall structure of Celerity
programs, the following measures are being taken:

- Compute task command groups must now by default capture everything by
  value. This is enforced through a static assertion. To disable this
  behavior, the CELERITY_STRICT_CGF_SAFETY macro can be set to 0.
- Buffers can now be copied.
- The call to buffer::get_access is now "const". This is a trade-off
  between retaining const-correctness and allowing CGFs to request
  accessors on by-value captured buffers without having to be marked as
  "mutable".
- An error message is emitted if the runtime detects a buffer going out
  of scope before all tasks have been completed.

Further improvements:

- While still only a single distr_queue can be created per process, it is
  now possible to copy it.
- Buffers can now also implicitly initialize the runtime, in case it
  hasn't been initialized explicitly and no distr_queue has been created
  before it.
- Runtime is destroyed once last buffer or distr_queue goes out of scope
  (instead of relying on static lifetime which isn't well-defined).
- All examples, which turned out to have been relying on undefined
  behavior until now (yikes!), are updated to the new API requirements.