Make the master node work
While the master node and worker nodes historically had certain
differences in their implementation (in particular regarding buffers),
that has not been the case for quite some time. For this reason we now
simply treat the master node the same as any other worker, except that
it additionally does all of the graph generation work.

However, since we currently cannot return placeholder host accessors for
master accesses during the pre-pass (because they don't exist in the
SYCL standard), some care has to be taken to avoid unnecessary data
transfers.