Improve logging behavior and configuration
This moves most of the existing logging to the "trace" level, making
Celerity's default output much less verbose. Additionally, the
CELERITY_LOG_LEVEL environment variable is improved to parse
human-readable log levels, such as "warn", "info" and so on.

This also updates spdlog to a newer version that supports configuration
of the color logging behavior. Currently we simply default to
auto-detecting TTY color support.