Fix unit tests, add some new ones
The change in 04278dad broke some assumptions in the graph generation
unit tests, requiring minor changes to various test cases. Additionally,
tests related to command graph generation are made a bit more resilient
against similar changes in the future, by explicitly checking the chunk
offsets generated for execution commands on different nodes.

This also adds test cases for the bug discovered in 80aa5233, as well as
for a related issue that was also surfaced within existing unit tests
through the change in 04278dad.