Improve automatic device selection strategy
Instead of defaulting to the first available device on each worker node,
Celerity will now try to find a platform that has enough devices
available such that it can assign a unique one to each node. This means
that setting CELERITY_DEVICES should no longer be necessary in most
situations.