Improve CI setup
The CI workflow now runs on the "push" event, instead of "pull_request".
This allows for commits in branches that are not part of a PR also to be
checked. The celerity/ci-report-action job has been adjusted to report
warnings as commit comments, instead of commenting on pull requests.

Additionally, the checkout step now also does a recursive submodule
checkout.