Get rid of nested SYCL command group submissions
Previously, in the live pass, a Celerity CGF was executed within a SYCL CGF. Buffer accesses done inside the Celerity CGF submitted further buffer copy operations to SYCL, so there were nested sycl::handler::submit calls. This was probably not conformant and broke with the new hipSYCL runtime.

This moves the Celerity CGF call out of the SYCL CGF, so that get_access calls can submit copy operations in a defined manner.