Introduce custom preprocessor flag for debugging
This introduces the CELERITY_DETAIL_ENABLE_DEBUG flag, which is set for
debug builds of the Celerity runtime, regardless of a user target's
build type. This is more reliable than using NDEBUG, as headers and
implementation files are always guaranteed to receive the flag, both
during library compilation and when linking to it.

Additionally:
   - Fix a bug within buffer_manager which caused it to have a different
     size for debug and non-debug builds (prompting this change).
   - Restrict Celerity installations to a single build configuration.
     This was effectively already the case (as we currently don't set
     CMAKE_<CONFIG>_POSTFIX), but now an error is generated whenever two
     or more build configurations are detected.