Don't read uninitialized memory in buffer_manager tests
Some of the tests for the `buffer_manager` try to determine whether
certain superfluous coherence transfers have NOT been performed. To do
so, these tests used to read uninitialized memory, checking that the
superflouous data was not there.

While this was never a great idea, the thinking was that it would be
highly unlikely to encounter the expected values in uninitialized
memory. While this has worked so far for dedicated GPUs, as it turns
out, this is not as unlikely for integrated GPUs that share RAM with the
CPU.

This change introduces a special test mode for the `buffer_manager`
that, when enabled, initializes all newly allocated buffers with a known
bit pattern. We can then check for this bit pattern to establish that no
coherence transfers have taken place.