Add support for SYCL 2020 accessor constructors
along with new enums for modes/targets and deduction from new tags.

A file for compatibility with ComputeCpp 2.5.0 was added since they do
not provide support for some of the features of SYCL 2020 we are using.

matmul, wave_sim and syncing examples were updated to the new accessor
constructor syntax.