Don't rely on CELERITY_TEST macro for instrumentation
The way this macro was used (set only from within unit tests, but not
for the compilation of the library itself) opens the door to all kinds
of ODR violations / UB issues. As none of the instrumentations currently
in place are on any performance critical path, we simply remove the
macro guards, only relying on checking a boolean flag instead.