Avoid creation of empty chunks
So far, tasks with a global size that was smaller than the number of
nodes would still be split into N chunks, one for each node, with some
chunks having an empty range. This introduces uneccesary overhead in
scheduling and execution of commands that are effectively no-ops.
Additionally, since we currently do not support accessing empty buffer
ranges (which is a separate issue), empty chunks with e.g. a one-to-one
mapper would result in a crash.