Update authoritative version of clang-format to 14
This thankfully is mostly a documentation change, with only a few
formatting changes.

Also change the `AllowShortFunctionsOnASingleLine` rule from `Inline` to
`true`. This rule seems to be extremely buggy, and clang-format's
behavior for `Inline` starting with version 13 does not make any sense.
Earlier versions also didn't treat the rule correctly (basically all
short functions were put on a single line, regardless of their position
in the code), which is what `true` is supposed to do (at least to my
understanding).