Add mechanism for manually running CI against HEAD revisions
This adds an input setting for the `workflow_dispatch` trigger (i.e.,
when manually launching a run through the GitHub UI) that determines
whether to run against the HEAD revisions, normally only tested during
nightly builds.

This is useful when fixing compatibility with HEAD revisions, so that
they can be tagged as "latest". A second new setting determines whether
the images should be tagged as well.