Remove reduction_manager from graph generation and printing
reduction_manager was so far used both for rid -> bid mapping and for storing / executing reductions themselves.
Replicating the rid -> bid mapping in graphs allows us to disentangle and simplify a lot of code, and also fixes a bug
where print_graph would fail on runtime shutdown because an already-finished reduction could not be resolved anymore.

In tests, reduction_manager is replaced by the new mock_reduction_factory.