clang-tidy: Enable (some) misc-* checks; do const-correctness pass
Clang-tidy 15 includes a new check "misc-const-correctness", which marks
locations where local variables could be qualified as const.
Unfortunately the check does not seem to be robust enough yet to be
enabled as an error (i.e., it includes some false positives).
For the time being it is also not checked during CI, as the CI
containers currently use Clang 14.

Nevertheless, a `clang-tidy --fix` pass has been applied to update many
locations to be qualified as const. A workaround has been added to
disable the diagnostic on several types (`buffer`, `accessor`,
`host_object`, `side_effect`) that could be marked const in most
situations, but shouldn't be from a semantic perspective.

Additionally, the clang-format configuration now prescribes the position
of CV-qualifiers to be to the left of types (i.e., "west-const") through
an option that was introduced with Clang 14.