*** BEGIN OF CCGRID23 UPSTREAMING PROJECT ***
The effort of upstreaming all the changes made for the CCGRID23 paper
will be split into several stages to make the process more transparent
and hopefully ensure that we can avoid performance regressions.

Each stage of the project will have two sides, a "left" side and a "right"
side. The left side starts from the current master branch and works its way
towards containing polished versions of all changes made for CCGRID23. The
right side contains the remainder of unpolished changes on top of the current
left side. In the end, the two sides converge to one final, fully integrated
branch.

** A message like this will be inserted as the final commit in each stage **

This branch is the left side of stage 0 and contains the state of
`master` as of 2023-01-17 (update: 2023-02-21).

Tentative plan for all stages:

[x] 0: Initial stage (no changes)
[ ] 1: USM buffers
[ ] 2: Remove multi-pass
[ ] 3: Buffer manager memory usage awareness
[ ] 4: Distributed scheduling
[ ] 5: Multi-GPU support
[ ] 6: ndv-buffers
[ ] 7: Async copies
[ ] 8: Pinned memory pool
[ ] X: Extra: Other and minor changes (may interleave with other stages)