Introduce backend system for vendor-specific code paths
Since SYCL 2020 does not support multi-dimensional (rectangular) copies
for USM pointers, we have to either do it in a loop (slow) or fall back
to vendor-specific APIs.

This introduces a new "backend" system that does the latter.
Currently only "generic" (= SYCL, slow) and CUDA (when using OpenSYCL or
DPC++) are supported.

Since backends are configuration during compile time, this additionally
introduces a new integration testing mechanism for testing backends.
This requires Celerity to be built with different CMake options, so the
test is implemented as a Python script.