Generate "fat" push commands
We now generate a single "fat" push command for each buffer and task,
instead of a separate push for each last writer and target node, similar
to how await pushes already work.

This will enable future optimizations in the IDAG, and also reduces the
number of commands that need to be generated, particularly in scenarios
where buffer contents are highly fragmented.