CI: Bump report action to v8
This fixes our build warning check that was broken since 0bbe6e564, due
to a version mismatch between the artifact upload action and the module
we use to download them inside the report action. If this breaks again
in the future, the action should now report this as a failure.

Additionally, a long standing issue where the report result would not
show up in CI runs triggered by a "pull_request" event is also fixed.