Measure and report executor starvation time
The executor is considered to be "starving" when it is out of
instructions to process, and the scheduler is currently busy.
This means that phases of the user program that do not interact with the
Celerity API do not count as starvation periods.

If the total starvation time exceeds a percentage threshold of the time
spent doing actual work (i.e., processing instructions), we print a
warning indicating that the application might be scheduler-bound.