[UPSTREAM][WIP] Optimize merge_overlapping_bounding_boxes for lookahead
Lookahead introduces a copy of each required box into the set of boxes
that need to be allocated for a task. If the number of boxes is large,
this can have a significant performance impact. We optimize this case by
not restarting the loop if the merged box was an exact copy.

TODO: Add a benchmark for this. The scenario can be derived from UMUGUC,
where the shape factor calculation can require the transfer of hundreds
or thousands of small subranges in a 1D buffer.